NPUW MoE Refactor: address review comments
